refactor: start with search page PR
removing modal things
