+++ packages/example-site/content/page/search.json
@@ -1,0 +1,5 @@
+{
+  "templateKey": "SearchPage",
+  "title": "Search the Whole Site!",
+  "image": "background.jpg"
+}
