+++ packages/gatsby-theme-project-portal/src/layouts/Layout.tsx
@@ -1,12 +1,6 @@
 import React, { FunctionComponent, ReactNode, useState } from "react"
 import { graphql } from "gatsby"
-import {
-  Footer,
-  BottomBanner,
-  DevelopmentBanner,
-  Navbar,
-  TopicType,
-} from "../components"
+import { Footer, BottomBanner, DevelopmentBanner, Navbar } from "../components"
 import { ImageDataLike } from "gatsby-plugin-image"
 import Modal from "react-modal"
 
@@ -45,61 +39,6 @@ export interface LayoutProps {
         }
       }
     }
-    allProject: {
-      nodes: {
-        slug
-        question
-        title
-        summary
-        status
-        opportunityCloses
-        startDate
-        endDate
-        lastModified
-        agency
-        topics: TopicType[]
-        deliverable
-        purpose
-        expertise
-        requirement
-        keyDates
-        priorResearch
-        statusOfData
-        fundingInfo
-        emailContent
-        mainContact: {
-          name
-          title
-          employer
-          email
-        }
-        projectTeam: {
-          name
-          title
-          employer
-          email
-        }
-        faq: {
-          text
-          title
-        }
-      }
-    }
-    allGeneralPage: {
-      nodes: {
-        slug
-        lede
-        faq: {
-          text
-          title
-        }
-        aims: {
-          text
-          title
-        }
-        title
-      }
-    }
     navbarLogo?: ImageDataLike
     bottomBannerLogo?: ImageDataLike
     footerLogo?: ImageDataLike
@@ -117,15 +56,12 @@ export const Layout: FunctionComponent<LayoutProps> = ({
       pages,
       staticText: { bottomBanner, footer },
     },
-    allProject,
-    allGeneralPage,
     navbarLogo,
     bottomBannerLogo,
     footerLogo,
   },
   children,
 }) => {
-  const searchNodes = { siteUrl, allProject, allGeneralPage }
   Modal.setAppElement(`#___gatsby`)
   return (
     <div className="w-full mx-0 bg-white border-0 xl:container xl:p-0 xl:mx-auto xl:border-l xl:border-r xl:border-gray-200 flex flex-col min-h-screen">
@@ -135,7 +71,6 @@ export const Layout: FunctionComponent<LayoutProps> = ({
         image={navbarLogo}
         pages={pages}
         activePage={path}
-        searchNodes={searchNodes}
       />
       <div className="flex-1">{children}</div>
       <BottomBanner
@@ -215,63 +150,6 @@ export const query = graphql`
     ) {
       childImageSharp {
         gatsbyImageData(height: 64)
-      }
-    }
-    allProject {
-      nodes {
-        slug
-        question
-        title
-        summary
-        status
-        opportunityCloses
-        startDate
-        endDate
-        lastModified
-        agency
-        topics {
-          ...TopicDetails
-        }
-        deliverable
-        purpose
-        expertise
-        requirement
-        keyDates
-        priorResearch
-        statusOfData
-        fundingInfo
-        emailContent
-        mainContact {
-          name
-          title
-          employer
-          email
-        }
-        projectTeam {
-          name
-          title
-          employer
-          email
-        }
-        faq {
-          text
-          title
-        }
-      }
-    }
-    allGeneralPage {
-      nodes {
-        slug
-        lede
-        faq {
-          text
-          title
-        }
-        aims {
-          text
-          title
-        }
-        title
       }
     }
   }
